Minor rename (#1267)
