+++ spring-cloud-kubernetes-fabric8-discovery/src/test/java/org/springframework/cloud/kubernetes/fabric8/discovery/Fabric8KubernetesDiscoveryClientTests.java
@@ -50,7 +50,7 @@ import org.springframework.cloud.kubernetes.commons.discovery.KubernetesDiscover
  */
 @EnableKubernetesMockClient(crud = true, https = false)
 @ExtendWith(OutputCaptureExtension.class)
-class KubernetesDiscoveryClientTests {
+class Fabric8KubernetesDiscoveryClientTests {
 
 	private static KubernetesClient client;
